revm_precompile::bls12_381::g2_msm

Function g2_msm

pub fn g2_msm(input: &[u8], gas_limit: u64) -> PrecompileResult

Implements EIP-2537 G2MSM precompile. G2 multi-scalar-multiplication call expects 288*k bytes as an input that is interpreted as byte concatenation of k slices each of them being a byte concatenation of encoding of G2 point (256 bytes) and encoding of a scalar value (32 bytes). Output is an encoding of multi-scalar-multiplication operation result - single G2 point (256 bytes). See also: https://eips.ethereum.org/EIPS/eip-2537#abi-for-g2-multiexponentiation