revm_precompile

Module utilities

Utility function that precompiles use, padding and converting between types.

Functions

bool_to_b256

Converts a boolean to a left-padded B256 value.

bool_to_bytes32

Converts a boolean to a left-padded 32-byte Bytes value.

left_pad

Left-pads the given slice with zeroes until LEN.

left_pad_vec

Left-pads the given slice with zeroes until len.

right_pad

Right-pads the given slice with zeroes until LEN.

right_pad_vec

Right-pads the given slice with zeroes until len.

right_pad_with_offset

Right-pads the given slice at offset with zeroes until LEN.

right_pad_with_offset_vec

Right-pads the given slice at offset with zeroes until len.